moved dga dataset class to analytics package.
