+++ python/clx/tests/test_dga_detector.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import cudf
-from clx.utils.data.dga_dataset import DGADataset
+from clx.analytics.dga_dataset import DGADataset
 from clx.utils.data.dataloader import DataLoader
 from clx.analytics.dga_detector import DGADetector
 from clx.analytics.model.rnn_classifier import RNNClassifier
